WIP
Fixed division and added a safe version that's not yet
been tested so is probably unsafe. Broke projection in
the process, and then realised some parts of the matrix
could be pre-calculated and simplified.

todo: work out what actually broke projection because
that didn't actually use the division macro in the
first place.